printk/nbcon: Remove the msleep() from nbcon_kthread_should_wakeup().
The msleep() in nbcon_kthread_should_wakeup() is harmfull as it destroy
the task state. This in turn leads to other problems such that the
printing threads can never rest.

It seems usefull during boot if the console is used by another thread
and instead of spinning it takes a small break instead of spinning.
However this is "brief" and if the printing thread is using the console
and the tty user is "blocked" then it is spinning so…

Remove the msleep() from nbcon_kthread_should_wakeup().

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>